Revert renaming of test data file
